DCS-64 Extracting out update builder
