+++ server/app.js
@@ -170,13 +170,26 @@ module.exports = function createApp({ signInService, incidentService, offenderSe
     app.use(csurf())
   }
 
-  // token refresh
+  // JWT token refresh
   app.use(async (req, res, next) => {
-    if (req.user) {
+    if (req.user && req.originalUrl !== '/logout') {
       const timeToRefresh = new Date() > req.user.refreshTime
       if (timeToRefresh) {
-        req.session.returnTo = req.originalUrl
-        return res.redirect('/login')
+        try {
+          const newToken = await signInService.getRefreshedToken(req.user)
+          req.user.token = newToken.token
+          req.user.refreshToken = newToken.refreshToken
+          logger.info(`existing refreshTime in the past by ${new Date() - req.user.refreshTime}`)
+          logger.info(
+            `updating time by ${newToken.refreshTime - req.user.refreshTime} from ${req.user.refreshTime} to ${
+              newToken.refreshTime
+            }`
+          )
+          req.user.refreshTime = newToken.refreshTime
+        } catch (error) {
+          logger.error(`Token refresh error: ${req.user.username}`, error.stack)
+          return res.redirect('/logout')
+        }
       }
     }
     return next()
@@ -225,7 +238,7 @@ module.exports = function createApp({ signInService, incidentService, offenderSe
   app.use(currentUserInContext)
 
   app.use('/', createIncidentsRouter({ authenticationMiddleware, incidentService, offenderService }))
-  app.use('/check-answers/', createCheckAnswersRouter({ authenticationMiddleware, incidentService, offenderService }))
+  app.use('/check-answers/', createCheckAnswersRouter({ authenticationMiddleware, incidentService }))
   app.use('/submitted/', createSubmittedRouter({ authenticationMiddleware }))
   app.use('/form/', createFormRouter({ authenticationMiddleware, incidentService, offenderService }))
   app.use('/api/', createApiRouter({ authenticationMiddleware, offenderService }))
