+++ server/services/incidentService.js
@@ -1,6 +1,6 @@
 const logger = require('../../log.js')
-const { equals } = require('../utils/utils')
 const { validate } = require('../utils/fieldValidation')
+const getUpdatedFormObject = require('./updateBuilder')
 
 module.exports = function createIncidentService({ formClient, elite2ClientBuilder }) {
   async function getFormResponse(userId, bookingId) {
@@ -33,46 +33,6 @@ module.exports = function createIncidentService({ formClient, elite2ClientBuilde
       })
     }
     return updatedFormObject
-  }
-
-  function getUpdatedFormObject({ formObject, fieldMap, userInput, formSection, formName }) {
-    const answers = fieldMap.reduce(answersFromMapReducer(userInput), {})
-
-    const updatedFormObject = {
-      ...formObject,
-      [formSection]: {
-        ...formObject[formSection],
-        [formName]: answers,
-      },
-    }
-    return !equals(formObject, updatedFormObject) && updatedFormObject
-  }
-
-  function answersFromMapReducer(userInput) {
-    return (answersAccumulator, field) => {
-      const { fieldName, answerIsRequired, sanitiser } = getFieldInfo(field, userInput)
-
-      if (!answerIsRequired) {
-        return answersAccumulator
-      }
-
-      return { ...answersAccumulator, [fieldName]: sanitiser(userInput[fieldName]) }
-    }
-  }
-
-  function getFieldInfo(field, userInput) {
-    const fieldName = Object.keys(field)[0]
-    const fieldConfig = field[fieldName]
-
-    const fieldDependentOn = userInput[fieldConfig.dependentOn]
-    const { sanitiser = value => value, predicate: predicateResponse } = fieldConfig
-    const dependentMatchesPredicate = fieldConfig.dependentOn && fieldDependentOn === predicateResponse
-
-    return {
-      fieldName,
-      sanitiser,
-      answerIsRequired: !fieldDependentOn || dependentMatchesPredicate,
-    }
   }
 
   const getIncidentsForUser = async (userId, status) => {
