+++ server/services/incidentService.test.js
@@ -43,114 +43,17 @@ describe('getFormResponse', () => {
   })
 })
 
-describe('update', () => {
-  const baseForm = {
-    section1: '',
-    section2: '',
-    section3: {},
-    section4: {
-      form1: {},
-      form2: { answer: 'answer' },
-    },
-  }
+describe('getIncidentsForUser', () => {
+  test('retrieve  details', async () => {
+    const output = await service.getIncidentsForUser('user1', 'STATUS-1')
 
-  describe('When there are no dependant fields', () => {
-    const fieldMap = [{ decision: {} }, { followUp1: {} }, { followUp2: {} }]
+    expect(output).toEqual([{ id: 1 }, { id: 2 }])
 
-    const form = {
-      ...baseForm,
-      section4: {
-        ...baseForm.section4,
-        form3: {
-          decision: '',
-          followUp1: '',
-          followUp2: '',
-        },
-      },
-    }
-
-    test('should build updated object correctly', async () => {
-      const userInput = {
-        decision: 'Yes',
-        followUp1: 'County',
-        followUp2: 'Town',
-      }
-
-      const output = await service.getUpdatedFormObject({
-        formObject: baseForm,
-        fieldMap,
-        userInput,
-        formSection: 'section4',
-        formName: 'form3',
-      })
-
-      expect(output).toEqual({
-        ...form,
-        section4: {
-          ...form.section4,
-          form3: {
-            decision: 'Yes',
-            followUp1: 'County',
-            followUp2: 'Town',
-          },
-        },
-      })
-    })
-
-    test('should not return updated form object when there has been no change', async () => {
-      const fieldMapSimple = [{ answer: {} }]
-      const userInput = { answer: 'answer' }
-
-      const existingForm = {
-        section1: '',
-        section2: '',
-        section3: {},
-        section4: {
-          form1: {},
-          form2: { answer: 'answer' },
-        },
-      }
-
-      const output = await service.getUpdatedFormObject({
-        formObject: existingForm,
-        fieldMap: fieldMapSimple,
-        userInput,
-        formSection: 'section4',
-        formName: 'form2',
-      })
-
-      expect(output).toEqual(false)
-    })
-
-    it('should add new sections and forms to the form if they dont exist', async () => {
-      const userInput = {
-        decision: 'Yes',
-        followUp1: 'County',
-        followUp2: 'Town',
-      }
-
-      const output = await service.getUpdatedFormObject({
-        formObject: baseForm,
-        fieldMap,
-        userInput,
-        formSection: 'section5',
-        formName: 'form1',
-      })
-
-      const expectedForm = {
-        ...baseForm,
-        section5: {
-          form1: {
-            decision: 'Yes',
-            followUp1: 'County',
-            followUp2: 'Town',
-          },
-        },
-      }
-      expect(output).toEqual(expectedForm)
-    })
+    expect(formClient.getIncidentsForUser).toBeCalledTimes(1)
+    expect(formClient.getIncidentsForUser).toBeCalledWith('user1', 'STATUS-1')
   })
-
+})
+describe('update', () => {
   test('should call update and pass in the form when form id is present', async () => {
     const updatedFormObject = { decision: 'Yes', followUp1: 'County', followUp2: 'Town' }
 
@@ -184,96 +87,6 @@ describe('update', () => {
       formResponse: updatedFormObject,
     })
   })
-
-  describe('When there are dependant fields', () => {
-    const form = {
-      ...baseForm,
-      section4: {
-        ...baseForm.section4,
-        form3: {
-          decision: '',
-          followUp1: '',
-          followUp2: '',
-        },
-      },
-    }
-
-    const fieldMap = [
-      { decision: {} },
-      {
-        followUp1: {
-          dependentOn: 'decision',
-          predicate: 'Yes',
-        },
-      },
-      {
-        followUp2: {
-          dependentOn: 'decision',
-          predicate: 'Yes',
-        },
-      },
-    ]
-
-    test('should store dependents if predicate matches', async () => {
-      const userInput = {
-        decision: 'Yes',
-        followUp1: 'County',
-        followUp2: 'Town',
-      }
-
-      const formSection = 'section4'
-      const formName = 'form3'
-
-      const output = await service.getUpdatedFormObject({
-        formObject: baseForm,
-        fieldMap,
-        userInput,
-        formSection,
-        formName,
-      })
-
-      expect(output).toEqual({
-        ...form,
-        section4: {
-          ...form.section4,
-          form3: {
-            decision: 'Yes',
-            followUp1: 'County',
-            followUp2: 'Town',
-          },
-        },
-      })
-    })
-
-    test('should remove dependents if predicate does not match', async () => {
-      const userInput = {
-        decision: 'No',
-        followUp1: 'County',
-        followUp2: 'Town',
-      }
-
-      const formSection = 'section4'
-      const formName = 'form3'
-
-      const output = await service.getUpdatedFormObject({
-        formObject: baseForm,
-        fieldMap,
-        userInput,
-        formSection,
-        formName,
-      })
-
-      expect(output).toEqual({
-        ...form,
-        section4: {
-          ...form.section4,
-          form3: {
-            decision: 'No',
-          },
-        },
-      })
-    })
-  })
 })
 
 describe('getValidationErrors', () => {
@@ -300,42 +113,5 @@ describe('getValidationErrors', () => {
     ${{ q1: 'No' }}  | ${dependantConfig} | ${[]}
   `('should return errors $expectedContent for form return', ({ formBody, formConfig, expectedOutput }) => {
     expect(service.getValidationErrors(formBody, formConfig)).toEqual(expectedOutput)
-  })
-
-  test('sanitisation', async () => {
-    const config = [
-      {
-        q1: {
-          responseType: 'requiredString',
-          validationMessage: 'Please give a full name',
-          sanitiser: val => val.toUpperCase(),
-        },
-      },
-    ]
-
-    const output = await service.getUpdatedFormObject({
-      formObject: {},
-      fieldMap: config,
-      userInput: { q1: 'aaaAAAaa' },
-      formSection: 'section4',
-      formName: 'form1',
-    })
-
-    expect(output).toEqual({
-      section4: {
-        form1: {
-          q1: 'AAAAAAAA',
-        },
-      },
-    })
-  })
-
-  test('getIncidentsForUser', async () => {
-    const output = await service.getIncidentsForUser('user1', 'STATUS-1')
-
-    expect(output).toEqual([{ id: 1 }, { id: 2 }])
-
-    expect(formClient.getIncidentsForUser).toBeCalledTimes(1)
-    expect(formClient.getIncidentsForUser).toBeCalledWith('user1', 'STATUS-1')
   })
 })
